requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
black==23.11.0
celery==5.3.4
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.2
coverage-badge==1.1.0
distlib==0.4.0
Django==5.1.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-yasg==1.21.7
execnet==2.1.1
factory-boy==3.3.0
Faker==20.1.0
filelock==3.19.1
flake8==6.1.0
identify==2.6.13
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.7.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==3.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-rerunfailures==12.0
pytest-xdist==3.3.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
